Middleware now returns a 'StackItem'
This structure represents a single compiled middleware stack, consisting
of the final handler and the context that is used for all middleware in
the stack.  This allows us to access the context that was used for the
middleware, a prerequisite for being able to set the URL parameters in a
way that the final handler can access.